fix: update lib/

+++ lib/driver/driver.h
@@ -1,8 +1,10 @@
 #ifndef DRIVER_H
 #define DRIVER_H
+
 #include <Arduino.h>
+
 #include "config.h"
 
-void driver(int valorX, int valorY);
+void driver(float valorX, float valorY);
 
 #endif
